requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
certifi==2023.5.7
click==8.1.5
colorama==0.4.6
dictdiffer==0.9.0
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.2
pydantic==2.0.2
pydantic-core==2.1.2
pydantic-extra-types==2.0.0
pydantic-settings==2.0.1
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.11.8
tortoise-orm==0.19.3
typing-extensions==4.7.1
ujson==5.8.0
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
